api/requirements.txt
Django==1.8
djangorestframework==3.1.2
